Отрисуй меня полностью (часть 2). Шестая проверка
