+++ js/create-previews.js
@@ -1,5 +1,3 @@
-import { bigPicture } from './big-picture.js';
-
 const similarImagesBlock = document.querySelector('.pictures');
 const similarImagesTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
@@ -12,38 +10,10 @@ const createPreviews = (images) => {
     imagesElement.querySelector('.picture__likes').textContent = likes;
     imagesElement.querySelector('.picture__comments').textContent = comments.length;
     imagesElement.querySelector('.picture__img').setAttribute('data-id', id);
-
-    imagesElement.addEventListener('click', () => {
-      const idBigPreview = imagesElement.querySelector('img').getAttribute('data-id');
-      const matchingPreviewBigPicture =  images.find((photo) => String(photo.id) === idBigPreview);
-      const socialComments = bigPicture.querySelector('.social__comments');
-      const templateComment = document.querySelector('#comment').content.querySelector('.social__comment');
-      const similarPictureComment = document.createDocumentFragment();
-
-      bigPicture.querySelector('.social__comments').innerHTML = '';
-
-      matchingPreviewBigPicture.comments.forEach(({ avatar, message, name }) => {
-        const commentsElement = templateComment.cloneNode(true);
-        commentsElement.querySelector('.social__picture').src = avatar;
-        commentsElement.querySelector('.social__picture').alt = name;
-        commentsElement.querySelector('.social__text').textContent = message;
-        similarPictureComment.appendChild(commentsElement);
-      });
-
-      socialComments.appendChild(similarPictureComment);
-
-      bigPicture.querySelector('.social__caption').textContent = matchingPreviewBigPicture.description;
-      bigPicture.querySelector('img').src = matchingPreviewBigPicture.url;
-      bigPicture.querySelector('.likes-count').textContent = matchingPreviewBigPicture.likes;
-
-      bigPicture.querySelector('.social__comment-count').classList.add('hidden');
-      bigPicture.querySelector('.comments-loader').classList.add('hidden');
-    });
-
     similarImagesFragment.appendChild(imagesElement);
   });
 
   similarImagesBlock.appendChild(similarImagesFragment);
 };
 
-export { createPreviews, similarImagesBlock };
+export { createPreviews };
